+++ src/data/data-store/DataStoreKeys.ts
@@ -1,6 +1,6 @@
 export const DataStoreKeys = {
     MODULES: "modules",
     NEWS: "news",
-    DATA_SUBMISSIONS: "calls", //TO DO
+    DATA_SUBMISSIONS: "data-submissions",
     SUBMISSIONS: "submissions",
 };
