+++ src/domain/entities/GlassSubmissions.ts
@@ -11,6 +11,6 @@ export interface GlassSubmissions {
     specimens: string[];
     status: string;
     submissionDate: Date;
-    call: string; //TO DO
+    dataSubmission: string;
     module: string;
 }
