--- src/domain/repositories/GlassCallRepository.ts
@@ -1,6 +1,0 @@
-import { FutureData } from "../entities/Future";
-import { GlassCall } from "../entities/GlassCallStatus";
-
-export interface GlassCallRepository {
-    getSpecificCall(module: string, orgUnit: string, period: number): FutureData<GlassCall[]>;
-}
