+++ src/domain/repositories/GlassDataSubmissionRepository.ts
@@ -1,0 +1,6 @@
+import { FutureData } from "../entities/Future";
+import { GlassDataSubmission } from "../entities/GlassDataSubmission";
+
+export interface GlassDataSubmissionsRepository {
+    getSpecificDataSubmission(module: string, orgUnit: string, period: number): FutureData<GlassDataSubmission[]>;
+}
