+++ src/webapp/components/current-data-submission/UploadsTable.tsx
@@ -17,7 +17,7 @@ export interface UploadsDataItemProps {
     specimens: string[];
     status: string;
     submissionDate: Date;
-    call: string; //TO DO
+    dataSubmission: string;
     module: string;
 }
