+++ src/webapp/components/current-data-submission/ValidationCard.tsx
