+++ src/webapp/components/current-data-submission/mock-tables-data.json
