+++ src/webapp/components/current-data-submission/mock-validations-list.json
