+++ src/webapp/components/current-data-submission/overview/StatusDetails.ts
