+++ src/webapp/components/data-submissions-history/DataSubmissionsTableBody.tsx
@@ -1,17 +1,17 @@
 import React from "react";
 import { TableBody, TableCell, TableRow } from "@material-ui/core";
 import styled from "styled-components";
-import { CallsHistoryItemProps } from "./CallsTable";
+import { DataSubmissionsHistoryItemProps } from "./DataSubmissionsTable";
 import ChevronRightIcon from "@material-ui/icons/ChevronRight";
 import { glassColors } from "../../pages/app/themes/dhis2.theme";
 import { StatusCapsule } from "./StatusCapsule";
 import { useHistory, useLocation } from "react-router-dom";
 
-export interface CallsTableBodyProps {
-    rows?: CallsHistoryItemProps[];
+export interface DataSubmissionsTableBodyProps {
+    rows?: DataSubmissionsHistoryItemProps[];
 }
 
-export const CallsTableBody: React.FC<CallsTableBodyProps> = ({ rows }) => {
+export const DataSubmissionsTableBody: React.FC<DataSubmissionsTableBodyProps> = ({ rows }) => {
     const history = useHistory();
     // TODO: remove the next two lines and create a global hook to get current module
     const location = useLocation().pathname.slice(1);
@@ -25,7 +25,7 @@ export const CallsTableBody: React.FC<CallsTableBodyProps> = ({ rows }) => {
         <>
             {rows && rows.length ? (
                 <StyledTableBody>
-                    {rows.map((row: CallsHistoryItemProps) => (
+                    {rows.map((row: DataSubmissionsHistoryItemProps) => (
                         <TableRow key={row.id} onClick={click}>
                             <TableCell>{row.year}</TableCell>
                             <TableCell>{row.open_status}</TableCell>
