+++ src/webapp/pages/Router.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { HashRouter, Route, Switch } from "react-router-dom";
-import { CallsHistoryPage } from "./calls-history/CallsHistoryPage";
+import { DataSubmissionsHistoryPage } from "./data-submissions-history/DataSubmissionsHistoryPage";
 import { CountryInformationPage } from "./country-information/CountryInformationPage";
-import { CurrentCallPage } from "./current-call/CurrentCallPage";
+import { CurrentDataSubmissionPage } from "./current-data-submission/CurrentDataSubmissionPage";
 import { DataSubmissionPage } from "./data-submission/DataSubmissionPage";
 import { FakeLandingPage } from "./landing/FakeLandingPage";
 import { UploadHistoryPage } from "./upload-history/UploadHistoryPage";
@@ -22,7 +22,7 @@ export const Router: React.FC = React.memo(() => {
                 {/* Default route */}
                 <Route
                     path="/current-data-submission/:module"
-                    render={({ match }) => <CurrentCallPage moduleName={match.params.module} />}
+                    render={({ match }) => <CurrentDataSubmissionPage moduleName={match.params.module} />}
                 />
                 <Route
                     path="/data-submission/:module"
@@ -30,7 +30,7 @@ export const Router: React.FC = React.memo(() => {
                 />
                 <Route
                     path="/data-submissions-history/:module"
-                    render={({ match }) => <CallsHistoryPage moduleName={match.params.module} />}
+                    render={({ match }) => <DataSubmissionsHistoryPage moduleName={match.params.module} />}
                 />
                 <Route
                     path="/upload-history/:module"
